Now using ajax to retrieve the feeds
Instead of relying on PHP for all backend, most of the work is done
with javascript. PHP is still used as a database interface and to
make remote jaxa requests (modern browsers have a "same origin"
restriciton for javascript which prevents it from making remote
requests).

The feeds are still not ordered in the list.